// src/common/enums/tool.enum.ts
export enum ToolNameEnum {
  SWAP = 'swap',
  ADD_LIQUIDITY = 'addLiquidity',
  WITHDRAW = 'withdraw',
  STAKE = 'stake',
  UNSTAKE = 'unstake',
  CLAIM_FEE = 'claimFee',
  CLAIM_EMISSION = 'claimEmission',
  CLAIM_ALL_REWARDS = 'claimAllRewards',
  LOCK_TOKENS = 'lockTokens',
  VOTE = 'vote',
  SET_LOCK_TO_RELAY = 'setLockToRelay',
  EXTEND_LOCK = 'extendLock',
  INCREASE_LOCK = 'increaseLock',
  MERGE_LOCKS = 'mergeLocks',
  TRANSFER_LOCK = 'transferLock',
  WITHDRAW_LOCK = 'withdrawLock',
  CLAIM_LOCK_REWARDS = 'claimLockRewards',
  ClAIM_VOTE_REWARDS = 'claimVotingRewards',
  RESET_LOCK = 'resetLock',
  POKE_LOCK = 'pokeLock',
}